pyproject.toml
[project]
name = "humankind"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "elevation>=1.1.3",
    "geojson>=3.2.0",
    "geopandas>=1.1.1",
    "matplotlib>=3.10.7",
    "numpy>=2.3.5",
    "pillow>=12.0.0",
    "pyproj>=3.7.2",
    "pyrosm>=0.6.2",
    "pytest>=9.0.1",
    "pyyaml>=6.0.3",
    "rasterio>=1.4.3",
    "scipy>=1.16.3",
    "shapely>=2.1.2",
    "srtm-py>=0.3.7",
    "topojson>=1.10",
]
